Add some joi validation

+++ api/src/controllers/fields.ts
@@ -1,5 +1,5 @@
 import { TYPES } from '@directus/constants';
-import { isDirectusError } from '@directus/errors';
+import { ForbiddenError, isDirectusError } from '@directus/errors';
 import type { Field, RawField, Type } from '@directus/types';
 import { Router } from 'express';
 import Joi from 'joi';
@@ -10,6 +10,7 @@ import { respond } from '../middleware/respond.js';
 import useCollection from '../middleware/use-collection.js';
 import { FieldsService } from '../services/fields.js';
 import asyncHandler from '../utils/async-handler.js';
+import { isSystemField } from '@directus/system-data';
 
 const router = Router();
 
@@ -117,6 +118,14 @@ router.post(
 	respond,
 );
 
+const systemUpdateSchema = Joi.object({
+	collection: Joi.string(),
+	field: Joi.string(),
+	schema: Joi.object({
+		is_indexed: Joi.bool(),
+	})
+});
+
 router.patch(
 	'/:collection',
 	validateCollection,
@@ -126,7 +135,15 @@ router.patch(
 			schema: req.schema,
 		});
 
-		if (Array.isArray(req.body) === false) {
+		if (Array.isArray(req.body)) {
+			for (const fieldData of req.body) {
+				if (isSystemField(fieldData['collection']!, fieldData['field']!)) {
+					const { error } = systemUpdateSchema.validate(fieldData);
+
+					if (error) throw new InvalidPayloadError({ reason: error.message });
+				}
+			}
+		} else {
 			throw new InvalidPayloadError({ reason: 'Submitted body has to be an array' });
 		}
 
@@ -176,10 +193,14 @@ router.patch(
 			schema: req.schema,
 		});
 
-		const { error } = updateSchema.validate(req.body);
+		if (isSystemField(req.params['collection']!, req.params['field']!)) {
+			const { error } = systemUpdateSchema.validate(req.body);
 
-		if (error) {
-			throw new InvalidPayloadError({ reason: error.message });
+			if (error) throw new InvalidPayloadError({ reason: error.message });
+		} else {
+			const { error } = updateSchema.validate(req.body);
+
+			if (error) throw new InvalidPayloadError({ reason: error.message });
 		}
 
 		const fieldData: RawField = req.body;
@@ -212,6 +233,10 @@ router.delete(
 			accountability: req.accountability,
 			schema: req.schema,
 		});
+
+		if (isSystemField(req.params['collection']!, req.params['field']!)) {
+			throw new ForbiddenError();
+		}
 
 		await service.deleteField(req.params['collection']!, req.params['field']!);
 		return next();
